Move some fixtures
