+++ tests/test_Qcaclulator.py
@@ -14,48 +14,6 @@ per_s_cm2 = (1 / u.second) / (u.centimeter ** 2)
 c = const.c
 
 
-# Define some fixtures for Qcalculator.
-@pytest.fixture(
-    params=[
-        (np.arange(1, 101), np.random.random(100), None),
-        (np.linspace(2.1, 2.5, 200), np.random.random(200), np.random.random(200)),
-        (
-            np.linspace(0.5, 1.5, 50),
-            np.random.random(50),
-            np.floor(2 * np.random.random(50)),
-        ),
-    ]
-)
-def test_spec(request):
-    """Wave and flux, mask examples."""
-    return request.param
-
-
-# 3 situations each variable, no unit. a unit. or a dimensionless unscaled unit.
-@pytest.fixture(params=[1, u.micron, u.dimensionless_unscaled])
-def wav_unit(request):
-    """Iterate some units on wavelength."""
-    return request.param
-
-
-@pytest.fixture(params=[1, per_s_cm2, u.dimensionless_unscaled])
-def flux_unit(request):
-    """Iterate some units on flux"""
-    return request.param
-
-
-@pytest.fixture(params=[1, u.dimensionless_unscaled])
-def trans_unit(request):
-    """Iterate some units on mask/transmission."""
-    return request.param
-
-
-@pytest.fixture(params=[True, False])
-def grad_flag(request):
-    """Gradient flag parameter."""
-    return request.param
-
-
 def test_rvprev_calc(test_spec, wav_unit, flux_unit, trans_unit):
     """Test that RVprec_calc can handle inputs as Quantities or unitless and returns a scalar Quantity."""
     wav = test_spec[0] * wav_unit
